requirements.txt
fastapi[standard]
motor
python-dotenv
pydantic
pymongo
